+++ wo/core/fileutils.py
@@ -344,14 +344,27 @@ class WOFileUtils():
                     continue
         return True
 
-    def writein(self, path, content):
+    def textwrite(self, path, content):
         """
-            Write content in path
+            Write content into a file
         """
         Log.debug(self, "Writing content in {0}".format(path))
         try:
             with open("{0}".format(path),
                       encoding='utf-8', mode='w') as final_file:
+                final_file.write('{0}'.format(content))
+        except IOError as e:
+            Log.debug(self, "{0}".format(e))
+            Log.error(self, "Unable to write content in {0}".format(path))
+
+    def textappend(self, path, content):
+        """
+            Append content to a file
+        """
+        Log.debug(self, "Writing content in {0}".format(path))
+        try:
+            with open("{0}".format(path),
+                      encoding='utf-8', mode='a') as final_file:
                 final_file.write('{0}'.format(content))
         except IOError as e:
             Log.debug(self, "{0}".format(e))